requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.4.5
django-health-check==3.17.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.0.4
idna==3.4
mysqlclient==2.1.1
PyJWT==2.6.0
pytz==2020.1
requests==2.28.1
sqlparse==0.3.1
tzdata==2022.7

urllib3==1.26.13
